refactor: rename dtos folder to input folder
