+++ apps/api/v2/src/modules/repositories/booking/booking-repository.service.ts
@@ -1,4 +1,4 @@
-import type { CreateBookingInput } from "@/modules/booking/dtos/create-booking";
+import type { CreateBookingInput } from "@/modules/booking/input/create-booking";
 import { PrismaReadService } from "@/modules/prisma/prisma-read.service";
 import { PrismaWriteService } from "@/modules/prisma/prisma-write.service";
 import { Injectable } from "@nestjs/common";
